main/labs/project/mlapi/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["James Winegar <jameswine@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
transformers = "^4.35.2"
torch = [
  {version = "^2.1.1", "platform" = "win32"},
  {version = "^2.1.1", "platform" = "linux"},
  {version = "^2.1.1", "platform" = "darwin",  markers = "platform_machine == 'x86_64'" },
  {version = "^2.1.1", "platform" = "darwin",  markers = "platform_machine == 'arm64'" },
  # {version = "^2.1.1+cu121", source = "pytorch", "platform" = "linux"}, # this is for training
]
requests = "^2.31"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.9"}
fastapi = "^0.104.1"
uvicorn = "^0.24"

[tool.poetry.group.dev.dependencies]
datasets = "^2.15"
scikit-learn = "^1.3.2"
pytest = "^7.4"
ruff = "^0.1"
black = "^23.11"
pre-commit = "^3.5"
httpx = "^0.25"
evaluate = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[[tool.poetry.source]]
name = "pypi"
priority = "primary"
